pyproject.toml
[project]
name = "pymodbus-gui"
version = "0.1.0"
description = "A GUI tool for Modbus communication using pymodbus"
authors = [
    { name = "clint", email = "clinton_luo@163.com" }
]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pymodbus==3.6.9",
    "pandas==2.2.2",
    "pyserial==3.5",
    "numpy==2.0.0",
    "PyQt6==6.7.0",
    "openpyxl (==3.1.2)",
]

[tool.poetry]
packages = [{ include = "pymodbus_gui", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
# 可选：开发依赖，比如 pytest、black 等
# pytest = "^8.0"
# black = "^24.0"